Split tmux session input tests
