+++ app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
@@ -42,7 +42,6 @@ import com.pocketshell.core.terminal.bridge.SshTerminalBridge
 import com.pocketshell.core.terminal.ui.TerminalSurfaceState
 import com.pocketshell.core.tmux.CommandResponse
 import com.pocketshell.core.tmux.TmuxClient
-import com.pocketshell.uikit.model.KeyKind
 import com.pocketshell.uikit.model.KeyModifierState
 import com.pocketshell.uikit.model.SessionAgentKind
 import com.pocketshell.core.tmux.TmuxClientException
@@ -51,7 +50,6 @@ import com.pocketshell.core.tmux.TmuxDisconnectEvent
 import com.pocketshell.core.tmux.TmuxDisconnectReason
 import com.pocketshell.core.tmux.TmuxOutputBacklogOverflow
 import com.pocketshell.core.tmux.protocol.ControlEvent
-import com.pocketshell.core.terminal.ui.TerminalRawInputPolicy
 import com.pocketshell.testsupport.drainMainLooperUntil as drainMainLooperUntilShared
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
@@ -7064,650 +7062,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
         assertTrue(
             "bridge-mode emulator must not generate OSC 11 color query responses, got $sent",
             sent.none { it.contains("send-keys -H") || it.contains("send-keys -l") },
-        )
-    }
-
-    @Test
-    fun terminalGeneratedInputResponseRoutesAsRawHex() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.writeInputToPane("%0", "\u001b]11;rgb:0101/0404/0909\u001b\\".toByteArray(Charsets.US_ASCII))
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(
-            listOf("send-keys -H -t %0 1b 5d 31 31 3b 72 67 62 3a 30 31 30 31 2f 30 34 30 34 2f 30 39 30 39 1b 5c"),
-            sent,
-        )
-    }
-
-    @Test
-    fun singleEscapeInputStillUsesNamedKeyPath() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.writeInputToPane("%0", byteArrayOf(0x1B))
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(listOf("send-keys -t %0 Escape"), sent)
-    }
-
-    @Test
-    fun writeInputToPaneIgnoresEmptyBytes() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.writeInputToPane("%0", ByteArray(0))
-        advanceUntilIdle()
-
-        assertTrue(
-            "empty input must not produce a send-keys command",
-            client.sentCommands.none { it.startsWith("send-keys") },
-        )
-    }
-
-    @Test
-    fun writeInputToPaneResultPropagatesFailedPaneWrite() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient().apply {
-            closeAndThrowOnCommandPrefix = "send-keys"
-            closeAndThrowException = TmuxClientException("failed to write tmux command `send-keys`")
-        }
-        vm.attachClientForTest(client)
-
-        val result = vm.writeInputToPaneResult("%0", "hello".toByteArray(Charsets.UTF_8))
-
-        assertTrue(result.isFailure)
-        val error = result.exceptionOrNull()
-        assertTrue("expected TmuxClientException, got ${error?.javaClass?.name}", error is TmuxClientException)
-        assertTrue(error?.message?.contains("send-keys") == true)
-        assertEquals(listOf("send-keys -l -t %0 -- 'hello'"), client.sentCommands.filter { it.startsWith("send-keys") })
-        assertTrue("failed pane write must close the dead tmux client", client.closed)
-    }
-
-    // ------------------------------------------------------------- Issue #209
-    // Bracketed-paste wrapping for multi-line input. Single-line input
-    // must keep the existing `send-keys -l` + `send-keys ... Enter` shape
-    // so we don't regress per-line named-key routing for normal typing.
-
-    @Test
-    fun writeInputToPaneWrapsMultiLineInputInBracketedPaste() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        val payload = "para one\npara two\npara three"
-        vm.writeInputToPane("%4", payload.toByteArray(Charsets.UTF_8))
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(
-            "expected bracketed paste start, body, and end commands, got $sent",
-            3,
-            sent.size,
-        )
-        val cmd = sent[1]
-        assertTrue(
-            "expected send-keys -H targeting %4, got '$cmd'",
-            cmd.startsWith("send-keys -H -t %4 "),
-        )
-        // The hex payload should begin with the bracketed-paste start
-        // marker bytes (\e[200~ -> 1b 5b 32 30 30 7e) and end with the
-        // matching end marker (\e[201~ -> 1b 5b 32 30 31 7e). The
-        // newlines inside the body must reach tmux as `0a` bytes (i.e.
-        // literal LF), NOT as separate `send-keys ... Enter` calls.
-        assertTrue(
-            "expected bracketed-paste start marker in hex payload, got '$cmd'",
-            sent.first().endsWith("1b 5b 32 30 30 7e"),
-        )
-        assertTrue(
-            "expected bracketed-paste end marker in hex payload, got '$cmd'",
-            sent.last().endsWith("1b 5b 32 30 31 7e"),
-        )
-        // Three paragraphs separated by two `\n` bytes inside the
-        // markers.
-        val hexBody = cmd.substringAfter("send-keys -H -t %4 ")
-        val newlineCount = hexBody.split(' ').count { it == "0a" }
-        assertEquals(
-            "expected exactly 2 literal LF bytes inside bracketed paste, got '$hexBody'",
-            2,
-            newlineCount,
-        )
-        // No standalone `send-keys ... Enter` should fire — the whole
-        // block went through `-H`.
-        assertTrue(
-            "multi-line input must not emit a separate Enter named-key, got $sent",
-            sent.none { it.contains(" Enter") },
-        )
-    }
-
-    @Test
-    fun writeInputToPaneNormalisesCrLfToLfInsideBracketedPaste() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        // Windows-style line endings should collapse to LF only — we
-        // never want two paragraph separators where the source had one.
-        vm.writeInputToPane("%0", "alpha\r\nbeta".toByteArray(Charsets.UTF_8))
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals("expected start, body, and end commands, got $sent", 3, sent.size)
-        val cmd = sent[1]
-        val hexBody = cmd.substringAfter("send-keys -H -t %0 ")
-        val tokens = hexBody.split(' ')
-        assertEquals(
-            "expected exactly 1 LF (not CR LF) inside the paste, got '$hexBody'",
-            1,
-            tokens.count { it == "0a" },
-        )
-        assertEquals(
-            "expected no CR bytes inside the paste, got '$hexBody'",
-            0,
-            tokens.count { it == "0d" },
-        )
-    }
-
-    @Test
-    fun writeInputToPaneKeepsSingleLineInputOnTheLiteralPath() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        // Single-line input must NOT be wrapped in bracketed-paste.
-        // The existing send-keys -l shape is preserved so the regression
-        // suite around named-key Enter routing (and the per-line Enter
-        // semantics of `writeInputToPaneIssuesSendKeysWithLiteralBytes`)
-        // keeps working.
-        vm.writeInputToPane("%0", "hello".toByteArray(Charsets.UTF_8))
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(1, sent.size)
-        assertEquals("send-keys -l -t %0 -- 'hello'", sent[0])
-        assertTrue(
-            "single-line input must not use the -H bracketed-paste path, got $sent",
-            sent.none { it.startsWith("send-keys -H") },
-        )
-    }
-
-    @Test
-    fun writeInputToPaneTrailingNewlineGoesThroughBracketedPaste() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        // "text\n" contains a `\n` so we route it through the paste path
-        // even though there is only one line of content. This preserves
-        // the design: any `\n` in the input means "treat as a paste".
-        // Submission of the input is the caller's responsibility (they
-        // can send a separate Enter named-key after the paste lands).
-        vm.writeInputToPane("%0", "ls\n".toByteArray(Charsets.UTF_8))
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(
-            "expected start, body, and end send-keys -H invocations for `\\n`-terminated input, got $sent",
-            3,
-            sent.size,
-        )
-        assertTrue(
-            "expected send-keys -H, got '$sent'",
-            sent.all { it.startsWith("send-keys -H -t %0 ") },
-        )
-    }
-
-    @Test
-    fun largeBracketedPasteIsSplitIntoBoundedSendKeysCommands() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        val payload = buildString {
-            append("first line\n")
-            repeat(TMUX_PASTE_BODY_CHUNK_BYTES * 3) { append(('a'.code + (it % 26)).toChar()) }
-        }
-        vm.writeInputToPane("%0", payload.toByteArray(Charsets.UTF_8))
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys -H") }
-        assertTrue("expected multiple bounded paste chunks, got ${sent.size}: $sent", sent.size > 3)
-        assertTrue("paste start marker must be its own bounded command", sent.first().endsWith("1b 5b 32 30 30 7e"))
-        assertTrue("paste end marker must be its own bounded command", sent.last().endsWith("1b 5b 32 30 31 7e"))
-        val maxHexTokens = sent.drop(1).dropLast(1)
-            .maxOf { command -> command.substringAfter("send-keys -H -t %0 ").split(' ').size }
-        assertTrue(
-            "body chunks must be bounded to $TMUX_PASTE_BODY_CHUNK_BYTES bytes; max tokens=$maxHexTokens",
-            maxHexTokens <= TMUX_PASTE_BODY_CHUNK_BYTES,
-        )
-        assertTrue(
-            "large paste must not fall back to one unbounded command",
-            sent.none { it.substringAfter("send-keys -H -t %0 ").split(' ').size > TMUX_PASTE_BODY_CHUNK_BYTES },
-        )
-    }
-
-    @Test
-    fun buildBracketedPasteHexEmitsExpectedSequenceForKnownInput() {
-        val vm = newVm()
-        // Body: "a\nb" -> 0x61 0x0a 0x62.
-        // Wrapped with the bracketed-paste markers:
-        //   1b 5b 32 30 30 7e   <- ESC [ 2 0 0 ~
-        //   61                  <- a
-        //   0a                  <- LF
-        //   62                  <- b
-        //   1b 5b 32 30 31 7e   <- ESC [ 2 0 1 ~
-        val hex = vm.buildBracketedPasteHexForTest("a\nb".toByteArray(Charsets.UTF_8))
-        assertEquals(
-            "1b 5b 32 30 30 7e 61 0a 62 1b 5b 32 30 31 7e",
-            hex,
-        )
-    }
-
-    @Test
-    fun containsLineBreakIsTrueOnlyForLf() {
-        val vm = newVm()
-        assertTrue(vm.containsLineBreakForTest("a\nb".toByteArray(Charsets.UTF_8)))
-        assertTrue(vm.containsLineBreakForTest("a\r\nb".toByteArray(Charsets.UTF_8)))
-        assertFalse(vm.containsLineBreakForTest("a b".toByteArray(Charsets.UTF_8)))
-        // A lone `\r` (rare on Android; carriage return without LF) is
-        // intentionally NOT treated as a paragraph break — the input-
-        // tokens path on the single-line route turns it into a tmux
-        // `Enter` named key, which is the right thing for shell prompts.
-        assertFalse(vm.containsLineBreakForTest("a\rb".toByteArray(Charsets.UTF_8)))
-        assertFalse(vm.containsLineBreakForTest(ByteArray(0)))
-    }
-
-    @Test
-    fun onKeyBarKeyTranslatesLabelsToTmuxNamedKeys() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.onKeyBarKey("%0", "Esc")
-        vm.onKeyBarKey("%0", "Tab")
-        // Issue #784: the panel's clean arrow glyphs (← ↑ ↓ →) map to the same
-        // tmux cursor-key named keys the old `‹ ⌃ ⌄ ›` did.
-        vm.onKeyBarKey("%0", "←")
-        vm.onKeyBarKey("%0", "↑")
-        vm.onKeyBarKey("%0", "↓")
-        vm.onKeyBarKey("%0", "→")
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(6, sent.size)
-        assertTrue(sent[0].endsWith("Escape"))
-        assertTrue(sent[1].endsWith("Tab"))
-        assertTrue(sent[2].endsWith("Left"))
-        assertTrue(sent[3].endsWith("Up"))
-        assertTrue(sent[4].endsWith("Down"))
-        assertTrue(sent[5].endsWith("Right"))
-        // All addressed to the right pane.
-        assertTrue(sent.all { it.contains("-t %0") })
-    }
-
-    /**
-     * Issue #893: the ⇧Tab (back-tab / Shift+Tab) hotkey must resolve, through
-     * the real [TmuxSessionViewModel.onKeyBarKey] mapping, to tmux's `BTab`
-     * named key — `send-keys -t <pane> BTab`. tmux emits the back-tab escape
-     * sequence `ESC [ Z` (0x1b 0x5b 0x5a) for `BTab`, which is what Claude Code
-     * listens for to cycle its permission/plan mode. This drives the real
-     * mapping (the panel's exact "⇧Tab" label), not a hand-rolled copy.
-     */
-    @Test
-    fun onKeyBarKeyShiftTabSendsTmuxBackTabNamedKey() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        // The label sent here is the EXACT label the KEYS panel renders for the
-        // back-tab key (TmuxHotkeyPanelSections), so this exercises the real
-        // panel→mapping contract.
-        vm.onKeyBarKey("%0", "⇧Tab")
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(
-            listOf("send-keys -t %0 BTab"),
-            sent,
-        )
-    }
-
-    @Test
-    fun onKeyBarKeyEnterSendsTmuxEnterNamedKeyWithoutReflow() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        // Issue #527: the dedicated Enter/Return key submits a newline/CR to
-        // the pane via the tmux named `Enter` key on the `send-keys` control
-        // channel. Both the glyph label and the legacy "Enter" alias map to
-        // the same sequence.
-        vm.onKeyBarKey("%0", "⏎")
-        vm.onKeyBarKey("%0", "Enter")
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(
-            listOf(
-                "send-keys -t %0 Enter",
-                "send-keys -t %0 Enter",
-            ),
-            sent,
-        )
-        // No resize/redraw path: the named-key route never uses
-        // `refresh-client` and never the literal `send-keys -l`/`-H` byte
-        // paths that would imply a paste/reflow.
-        assertTrue(
-            "Enter must not trigger a resize/refresh-client",
-            client.sentCommands.none { it.startsWith("refresh-client") },
-        )
-    }
-
-    @Test
-    fun onKeyBarKeySendsCuratedCtrlCombosAsRawControlBytes() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        // Issue #784: the hotkeys panel's Ctrl combos are DIRECT buttons (no
-        // lone-Ctrl modifier to arm). Each `^X` label maps to its control byte
-        // via the `send-keys -H` overlay path (no resize/redraw). `^B` (0x02,
-        // tmux prefix / Claude "ctrl-b ctrl-b", #677) is in the set. The legacy
-        // `Ctrl-C` / `Ctrl-D` aliases remain accepted for the byte mapping.
-        vm.onKeyBarKey("%0", "^A")
-        vm.onKeyBarKey("%0", "^B")
-        vm.onKeyBarKey("%0", "^C")
-        vm.onKeyBarKey("%0", "^D")
-        vm.onKeyBarKey("%0", "^E")
-        vm.onKeyBarKey("%0", "^L")
-        vm.onKeyBarKey("%0", "^R")
-        vm.onKeyBarKey("%0", "^Z")
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(
-            listOf(
-                "send-keys -H -t %0 01",
-                "send-keys -H -t %0 02",
-                "send-keys -H -t %0 03",
-                "send-keys -H -t %0 04",
-                "send-keys -H -t %0 05",
-                "send-keys -H -t %0 0c",
-                "send-keys -H -t %0 12",
-                "send-keys -H -t %0 1a",
-            ),
-            sent,
-        )
-    }
-
-    @Test
-    fun ctrlBHotkeyDoubleTapSendsRawCtrlBByteTwice() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        // Issue #677/#784: `^B` is a direct hotkey routed through `onKeyBarKey`
-        // and sends the raw Ctrl-B byte (0x02). Two rapid taps (Claude Code's
-        // "ctrl-b ctrl-b to background") each fire independently, so the pane
-        // receives `02` twice with no de-dup/throttle swallowing the second.
-        vm.onKeyBarKey("%0", "^B")
-        vm.onKeyBarKey("%0", "^B")
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(
-            listOf(
-                "send-keys -H -t %0 02",
-                "send-keys -H -t %0 02",
-            ),
-            sent,
-        )
-    }
-
-    @Test
-    fun hotkeyPanelSectionsAreDeDupedAndCarryTheAuditedSet() {
-        // Issue #784: the dedicated panel set — no duplicate `/`, `^B` present,
-        // clean arrow glyphs. Issue #787 added the INTERRUPT / EOF doubled
-        // chords (`^C×2` / `^D×2`), re-homed from the deleted palette. Issue
-        // #1091 FILLED the missing nano/TUI control keys into CTRL COMBOS
-        // (`^G ^J ^K ^O ^T ^U ^W ^X ^\`) and DELIBERATELY restored a sticky
-        // `Ctrl` modifier plus the a–z LETTERS grid so `Ctrl+<any letter>` is
-        // reachable. Issue #1332 reordered ARROWS to the TOP and introduced
-        // progressive disclosure: a compact COMMON set (arrows + Esc/Tab/Enter/
-        // ^C/^D) shown by default, everything else EXTENDED behind the expander.
-        // Verify the curated set is exactly what we expect in the new order.
-        val labels = TmuxHotkeyPanelSections.flatMap { it.keys }.map { it.label }
-        assertEquals(
-            listOf(
-                // ARROWS — Issue #1332: now the FIRST/top section.
-                "←", "↑", "↓", "→",
-                // COMMON — Issue #1332: the everyday essentials shown by default.
-                "Esc", "Tab", "Enter", "^C", "^D",
-                // CTRL COMBOS — Issue #1091 ordered by control byte; the
-                // previously-missing nano/TUI keys are filled in. Now EXTENDED.
-                "^A", "^B", "^C", "^D", "^E", "^G", "^J", "^K", "^L", "^O",
-                "^R", "^T", "^U", "^W", "^X", "^Z", "^\\",
-                // MORE KEYS — Issue #893 ⇧Tab (back-tab), Issue #1332 EXTENDED.
-                "⇧Tab",
-                // INTERRUPT / EOF — Issue #787 doubled chords. Now EXTENDED.
-                TmuxHotkeyInterruptX2Label, TmuxHotkeyEofX2Label,
-                // CTRL + LETTER — Issue #1091 sticky modifier. Now EXTENDED.
-                TmuxHotkeyCtrlModifierLabel,
-                // LETTERS — Issue #1091 a–z grid for Ctrl composition. Now EXTENDED.
-                "a", "b", "c", "d", "e", "f", "g", "h", "i", "j", "k", "l", "m",
-                "n", "o", "p", "q", "r", "s", "t", "u", "v", "w", "x", "y", "z",
-            ),
-            labels,
-        )
-        // Issue #893: the back-tab key is present (now in the EXTENDED MORE KEYS).
-        assertTrue(labels.contains("⇧Tab"))
-        // Issue #1332: the ONLY intentional duplicates are `^C` / `^D` — surfaced
-        // both in the common essentials row AND kept in the full CTRL COMBOS grid
-        // (the most-reached-for control keys). No OTHER label repeats (the
-        // maintainer's "/ appears twice" / "Esc duplicated" complaints still hold).
-        val duplicated = labels.groupingBy { it }.eachCount().filterValues { it > 1 }.keys
-        assertEquals(setOf("^C", "^D"), duplicated)
-        // Issue #1091: the sticky `Ctrl` modifier is now INTENTIONALLY present
-        // (it is the general `Ctrl+<letter>` escape hatch), and it is a
-        // Modifier-kind key, not a lone Regular button.
-        assertTrue(labels.contains(TmuxHotkeyCtrlModifierLabel))
-        val ctrlKey = TmuxHotkeyPanelSections.flatMap { it.keys }
-            .single { it.label == TmuxHotkeyCtrlModifierLabel }
-        assertEquals(KeyKind.Modifier, ctrlKey.kind)
-        // No `/` key in the panel.
-        assertFalse(labels.contains("/"))
-        // ^B (tmux prefix) restored.
-        assertTrue(labels.contains("^B"))
-        // Issue #1091: the filled nano/TUI control keys are present.
-        listOf("^G", "^J", "^K", "^O", "^T", "^U", "^W", "^X", "^\\").forEach {
-            assertTrue("CTRL COMBOS must offer the filled key $it", labels.contains(it))
-        }
-        // Issue #1091: the full a–z LETTERS grid is present for Ctrl composition.
-        ('a'..'z').forEach { c ->
-            assertTrue("LETTERS must offer '$c'", labels.contains(c.toString()))
-        }
-        // Issue #787: the doubled interrupt/EOF chords are present and DISTINCT
-        // from the single `^C`/`^D` (they're not aliases of the same label).
-        assertTrue(labels.contains(TmuxHotkeyInterruptX2Label))
-        assertTrue(labels.contains(TmuxHotkeyEofX2Label))
-        assertTrue(labels.contains("^C"))
-        assertTrue(labels.contains("^D"))
-        // Issue #1332: the arrow section is now the FIRST section (was last), still
-        // using the clean glyphs, marked as Arrow kind.
-        val arrows = TmuxHotkeyPanelSections.first().keys
-        assertEquals(listOf("←", "↑", "↓", "→"), arrows.map { it.label })
-        assertTrue(arrows.all { it.kind == KeyKind.Arrow })
-    }
-
-    @Test
-    fun onKeyBarKeyInterruptAndEofDoubledChordsSendByteTwice() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        // Issue #787: the re-homed `^C×2` / `^D×2` hotkeys are DISTINCT from the
-        // single `^C`/`^D` — they send the control byte TWICE (`repeatCount = 2`)
-        // so they actually interrupt the running agent / send EOF (Claude Code
-        // and many REPLs treat the first press as "press again to interrupt/exit").
-        vm.onKeyBarKey("%0", TmuxHotkeyInterruptX2Label)
-        vm.onKeyBarKey("%0", TmuxHotkeyEofX2Label)
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        // `sendControlInputToPane(..., repeatCount = 2)` emits one `send-keys -H`
-        // carrying the byte twice (03 03 for Ctrl-C, 04 04 for Ctrl-D).
-        assertEquals(
-            listOf(
-                "send-keys -H -t %0 03 03",
-                "send-keys -H -t %0 04 04",
-            ),
-            sent,
-        )
-    }
-
-    @Test
-    fun keyBarControlEscapeAndHotkeysClearSmartTextBeforeTmuxRawSends() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "\$0", "shell", paneIndex = 0)),
-        )
-        advanceUntilIdle()
-        val state = vm.panes.value.single().terminalState
-        val policies = mutableListOf<TerminalRawInputPolicy>()
-        state.setSmartTextStagingBridgeForTest { policy ->
-            policies += policy
-            if (policy == TerminalRawInputPolicy.FlushSmartText) {
-                client.sentCommands.add("flush-staged")
-            }
-        }
-
-        vm.onKeyBarKey("%0", "^C")
-        vm.onKeyBarKey("%0", "Esc")
-        vm.onKeyBarKey("%0", "→")
-        advanceUntilIdle()
-
-        assertEquals(
-            listOf(
-                TerminalRawInputPolicy.ClearSmartText,
-                TerminalRawInputPolicy.ClearSmartText,
-                TerminalRawInputPolicy.ClearSmartText,
-            ),
-            policies,
-        )
-        assertEquals(
-            listOf(
-                "send-keys -H -t %0 03",
-                "send-keys -t %0 Escape",
-                "send-keys -t %0 Right",
-            ),
-            client.sentCommands.filter { it.startsWith("send-keys") },
-        )
-        assertFalse(client.sentCommands.contains("flush-staged"))
-    }
-
-    @Test
-    fun keyBarEnterFlushesSmartTextBeforeTmuxEnter() = runTest(scheduler) {
-        val vm = newVm()
-        val literalFlushGate = CompletableDeferred<Unit>()
-        val client = FakeTmuxClient().apply {
-            sendCommandGatePrefix = "send-keys -l -t %0 -- 'staged'"
-            sendCommandGate = literalFlushGate
-        }
-        vm.attachClientForTest(client)
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "\$0", "shell", paneIndex = 0)),
-        )
-        advanceUntilIdle()
-        val state = vm.panes.value.single().terminalState
-        val policies = mutableListOf<TerminalRawInputPolicy>()
-        state.setSmartTextStagingBridgeForTest { policy ->
-            policies += policy
-            if (policy == TerminalRawInputPolicy.FlushSmartText) {
-                state.writeInput("staged".toByteArray(Charsets.UTF_8))
-            }
-        }
-
-        try {
-            vm.onKeyBarKey("%0", "⏎")
-            waitForSentCommandCount(client, expectedCount = 1)
-
-            assertEquals(listOf(TerminalRawInputPolicy.FlushSmartText), policies)
-            assertEquals(
-                "Enter must wait behind the queued SmartText flush while the literal send is suspended",
-                listOf("send-keys -l -t %0 -- 'staged'"),
-                client.sentCommands.filter { it.startsWith("send-keys") },
-            )
-
-            literalFlushGate.complete(Unit)
-            waitForSentCommandCount(client, expectedCount = 2)
-
-            assertEquals(
-                listOf(
-                    "send-keys -l -t %0 -- 'staged'",
-                    "send-keys -t %0 Enter",
-                ),
-                client.sentCommands.filter { it.startsWith("send-keys") },
-            )
-        } finally {
-            literalFlushGate.complete(Unit)
-            state.setSmartTextStagingBridgeForTest(null)
-        }
-    }
-
-    @Test
-    fun directTmuxControlHotkeyClearsSmartTextBeforeSendingRawBytes() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "\$0", "shell", paneIndex = 0)),
-        )
-        advanceUntilIdle()
-        val state = vm.panes.value.single().terminalState
-        val policies = mutableListOf<TerminalRawInputPolicy>()
-        state.setSmartTextStagingBridgeForTest { policy ->
-            policies += policy
-            if (policy == TerminalRawInputPolicy.FlushSmartText) {
-                client.sentCommands.add("flush-staged")
-            }
-        }
-
-        vm.sendControlInputToPane("%0", CtrlCByte, repeatCount = 2)
-        advanceUntilIdle()
-
-        assertEquals(listOf(TerminalRawInputPolicy.ClearSmartText), policies)
-        assertEquals(
-            listOf("send-keys -H -t %0 03 03"),
-            client.sentCommands.filter { it.startsWith("send-keys") },
-        )
-        assertFalse(client.sentCommands.contains("flush-staged"))
-    }
-
-    @Test
-    fun sendControlInputToPaneCanSendDoublePressPayload() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        vm.attachClientForTest(client)
-
-        vm.sendControlInputToPane("%0", CtrlCByte, repeatCount = 2)
-        vm.sendControlInputToPane("%0", CtrlDByte, repeatCount = 2)
-        advanceUntilIdle()
-
-        val sent = client.sentCommands.filter { it.startsWith("send-keys") }
-        assertEquals(
-            listOf(
-                "send-keys -H -t %0 03 03",
-                "send-keys -H -t %0 04 04",
-            ),
-            sent,
         )
     }
